Move specific vue JS modules in the vue directory
