+++ tests/js/modules/SearchTokenizer/SearchInput.test.js
@@ -32,7 +32,7 @@
 
 /* global GLPI */
 
-import SearchInput from "../../../../js/modules/SearchTokenizer/SearchInput.js";
+import SearchInput from "../../../../js/src/vue/Kanban/SearchInput.js";
 
 describe('Search Tokenizer Input', () => {
